pyproject.toml
[tool.poetry]
name = "verizon-drift-monitoring-poc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "verizon_drift_monitoring_poc"}]

[tool.poetry.dependencies]
python = "~3.10"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
numpy = "^1.25.0"
torchmetrics = "^0.11.4"
sqlalchemy = "^2.0.17"
pyspark = "^3.4.1"
matplotlib = "^3.7.1"
pyyaml = "^6.0"
requests = "^2.31.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
sqlalchemy-utils = "^0.41.1"
pyarrow = "^12.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
